Tried some ray casting stuff, made the level messy
I wanted to try out selecting an object by clicking on it, so I implemented shooting a ray from the controller on a trigger press and destroying a cube if it hits one. That worked, so now I'm messing with rotating an object after it's clicked (like a panel on the machine). Right now you can click cubes and they will spin like crazy lol