Light Scenario Functional
I worked up a light scenario. There are a number of emergency buttons and a emergency light. If all buttons are green (not pushed) the light will remain green. If one or more buttons are pushed the light will be red. Also cleaned up the scene and added more pick ups and other tweaks